+++ libsrc/grabber/CMakeLists.txt
@@ -1,4 +1,3 @@
-
 if (ENABLE_DISPMANX)
 	add_subdirectory(dispmanx)
 endif (ENABLE_DISPMANX)
@@ -6,3 +5,7 @@ endif (ENABLE_DISPMANX)
 if (ENABLE_V4L2)
 	add_subdirectory(v4l2)
 endif (ENABLE_V4L2)
+
+if (ENABLE_X11)
+	add_subdirectory(x11)
+endif()
