+++ libsrc/protoserver/CMakeLists.txt
@@ -10,7 +10,9 @@ include_directories(
 # Group the headers that go through the MOC compiler
 set(ProtoServer_QT_HEADERS
 		${CURRENT_HEADER_DIR}/ProtoServer.h
+		${CURRENT_HEADER_DIR}/ProtoConnection.h
 		${CURRENT_SOURCE_DIR}/ProtoClientConnection.h
+		${CURRENT_HEADER_DIR}/ProtoConnectionWrapper.h
 )
 
 set(ProtoServer_HEADERS
@@ -19,6 +21,8 @@ set(ProtoServer_HEADERS
 set(ProtoServer_SOURCES
 		${CURRENT_SOURCE_DIR}/ProtoServer.cpp
 		${CURRENT_SOURCE_DIR}/ProtoClientConnection.cpp
+		${CURRENT_SOURCE_DIR}/ProtoConnection.cpp
+		${CURRENT_SOURCE_DIR}/ProtoConnectionWrapper.cpp
 )
 
 set(ProtoServer_PROTOS
@@ -44,5 +48,5 @@ add_library(protoserver
 target_link_libraries(protoserver
 		hyperion
 		hyperion-utils
-        ${PROTOBUF_LIBRARIES}
-        ${QT_LIBRARIES})
+		${PROTOBUF_LIBRARIES}
+		${QT_LIBRARIES})
