+++ src/hyperion-x11/X11Wrapper.h
@@ -3,37 +3,37 @@
 #include <QTimer>
 
 // Hyperion-X11 includes
-#include "X11Grabber.h"
+#include <grabber/X11Grabber.h>
 
 class X11Wrapper : public QObject
 {
-	Q_OBJECT
+    Q_OBJECT
 public:
-	X11Wrapper(const unsigned cropHorizontal, const unsigned cropVertical, const unsigned pixelDecimation);
+    X11Wrapper(const unsigned cropHorizontal, const unsigned cropVertical, const unsigned pixelDecimation);
 
-	const Image<ColorRgb> & getScreenshot();
+    const Image<ColorRgb> & getScreenshot();
 
-	///
-	/// Starts the timed capturing of screenshots
-	///
-	void start();
+    ///
+    /// Starts the timed capturing of screenshots
+    ///
+    void start();
 
-	void stop();
+    void stop();
 
 signals:
-	void sig_screenshot(const Image<ColorRgb> & screenshot);
+    void sig_screenshot(const Image<ColorRgb> & screenshot);
 
 private slots:
-	///
-	/// Performs a single screenshot capture and publishes the capture screenshot on the screenshot
-	/// signal.
-	///
-	void capture();
+    ///
+    /// Performs a single screenshot capture and publishes the capture screenshot on the screenshot
+    /// signal.
+    ///
+    void capture();
 
 private:
-	/// The QT timer to generate capture-publish events
-	QTimer _timer;
+    /// The QT timer to generate capture-publish events
+    QTimer _timer;
 
-	/// The grabber for creating screenshots
-	X11Grabber _grabber;
+    /// The grabber for creating screenshots
+    X11Grabber _grabber;
 };
